requirements.txt
boto3>=1.28.0
snowflake-connector-python>=3.0.0
pandas>=2.0.0
python-dotenv>=1.0.0
pyarrow>=12.0.0
pyyaml>=6.0
